Make -DENABLE_FP8 conditional on CMake option
The definition `-DENABLE_FP8` is added only conditional on the used CUDA version. There is an additional CMake option `ENABLE_FP8` which is supposed to control whether FP8 is used. The option only controlled the message, but not the actual definition.

This change makes the compile option dependent on the CMake option like the message suggests it should be.